% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{biomass2c}
\alias{biomass2c}
\title{Calculates biomass conversion to carbon}
\usage{
biomass2c(biomass, method, ci, type, biome)
}
\arguments{
\item{biomass}{(usually kg or metric tonnes)}

\item{method}{Thomas_sim = simple with biomass 0.483 and 95\% ci of 0.003, can be used for error progression (default)
Matthews = simpliest with biomass:carbon 0.5 (Matthews 1993)
IPCC_sim = simple with biomass 0.477 (IPCC 2006)
IPCC = lookup by type (Angiosperm or Conifer) and biome (tropical, Subtropical/Mediterranean,Temperate/Boreal,all)
Thomas = lookup by type (Angiosperm or Conifer) and biome (tropical, Subtropical/Mediterranean,Temperate/Boreal,all)}

\item{ci}{confidence interval returns TRUE or False (default)}

\item{type}{Angiosperm or Conifer}

\item{biome}{tropical, Subtropical/Mediterranean,Temperate/Boreal,all}
}
\value{
either carbon value or list of carbon value with sd error
}
\description{
Calculates biomass conversion to carbon using multiple methods/citations
Mathews: simpliest giving 0.5
IPCC simple: 0.477
IPCC by type and biome
Mathews simple: 0.483 with 95\% confidence interval 0.003
Mathews by type and biome
}
\references{
Thomas, Sean C., and Adam R. Martin. "Carbon content of tree tissues: a synthesis." Forests 3.2 (2012): 332-352. https://www.mdpi.com/1999-4907/3/2/332
IPCC. Forest lands. Intergovernmental Panel on Climate Change Guidelines for National Greenhouse Gas Inventories; Institute for Global Environmental Strategies (IGES): Hayama,Japan, 2006; Volume 4, p. 83.
Matthews, G.A.R. (1993) The Carbon Content of Trees. Forestry Commission Technical Paper 4. Forestry Commission, Edinburgh. 21pp. ISBN: 0-85538-317-8
}
\author{
Justin Moat. J.Moat@kew.org
}
