% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_agc}
\alias{fc_agc}
\title{tariff number from vol and area}
\usage{
fc_agc(spcode, dbh, height, returnv = "AGC")
}
\arguments{
\item{spcode}{species code}

\item{dbh}{in cm}

\item{height}{in m}

\item{returnv}{what to return, Either AGC or All}
}
\value{
either AGC in tonnes \link{default}
or All list of tariff number, merchantable volume m3, stem volume m3, stem biomass (tonnes), stem carbon (tonnes), canopy carbon (tonnes) and root carbon (tonnes)
}
\description{
todo*
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org
}
