% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_broad_sap_seedling2C}
\alias{fc_broad_sap_seedling2C}
\title{Broadleaf seedlings and saplings to carbon}
\usage{
fc_broad_sap_seedling2C(heightincm)
}
\arguments{
\item{height}{= tree height in metres}
}
\value{
carbon in tonnes
}
\description{
Broadleaf seedlings and saplings to carbon
}
\note{
just uses simple linear relationship to get between measures
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018)
}
\author{
Justin Moat. J.Moat@kew.org
}
