% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_con_sap_seedling2C}
\alias{fc_con_sap_seedling2C}
\title{Conifer seedlings and saplings to carbon}
\usage{
fc_con_sap_seedling2C(heightincm)
}
\arguments{
\item{heightincm}{tree height in metres}
}
\value{
carbon in tonnes
}
\description{
Conifer seedlings and saplings to carbon
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018)
}
\author{
Justin Moat. J.Moat@kew.org
}
