% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_con_tariff}
\alias{fc_con_tariff}
\title{Conifer tree tariff number}
\usage{
fc_con_tariff(spcode, height, dbh)
}
\arguments{
\item{height}{tree height in m}

\item{dbh}{diameter at breast height in cm}
}
\value{
tariff number
}
\description{
Use DBH and tree height to calculate the tariff number of each sample tree. Rounded to the nearest whole number. Species-specific estimates of a1 – a3 are found in the R data file, 'tariff_coniferdf'.
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org
}
