% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_rootbiomass}
\alias{fc_rootbiomass}
\title{Forestry commission root biomass estimates}
\usage{
fc_rootbiomass(spcode, dbh)
}
\arguments{
\item{spcode}{species code}

\item{dbh}{in cm}
}
\value{
biomass (oven dry tonnes)
}
\description{
todo*
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org
}
