% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_stand_tariff}
\alias{fc_stand_tariff}
\title{tariff number by stand height}
\usage{
fc_stand_tariff(spcode, height)
}
\arguments{
\item{spcode}{species code}

\item{height}{tree height in metres}
}
\value{
tariff number
}
\description{
tariff number by stand height
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org
}
