% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_tariff_vol_area}
\alias{fc_tariff_vol_area}
\title{tariff number from vol and area}
\usage{
fc_tariff_vol_area(vol, dbh)
}
\arguments{
\item{vol}{= tree volume in m^3}

\item{dbh}{in cm}
}
\value{
tariff number
}
\description{
tariff number from vol and area
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
(Equation 1)
}
\author{
Justin Moat. J.Moat@kew.org
}
