% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_treevol}
\alias{fc_treevol}
\title{Forestry commission tree wood volume}
\usage{
fc_treevol(mtreevol, dbh)
}
\arguments{
\item{mtreevol}{= mercantable tree volume}
}
\value{
volume m3
}
\description{
Forestry commission tree wood volume
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org
}
