% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_woodbiomass}
\alias{fc_woodbiomass}
\title{Forestry commission wood biomass}
\usage{
fc_woodbiomass(treevol, nsg)
}
\arguments{
\item{treevol}{tree volume m^3}

\item{nsg}{Nominal Specific Gravity (NSG)}
}
\value{
biomass (oven dry tonnes)
}
\description{
Forestry commission wood biomass
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org
}
