% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pro_error_carbon}
\alias{pro_error_carbon}
\title{Carbon progression of errors}
\usage{
pro_error_carbon(
  vol,
  volsd,
  den,
  densd,
  biom,
  biomsd,
  nruns = 10000,
  returnsv = NULL
)
}
\arguments{
\item{volsd}{volume sd}

\item{densd}{wood density sd}

\item{biomsd}{biomass sd}

\item{nruns}{number of iteration, suggest 10,000 as min and 100,000 is a good number}

\item{returnv}{if null then mean and sd is returned else vector of quantiles ie c(5,50,95)/100 will return 5\%, mean and 95\% quantiles.}
}
\value{
either vector of mean and sd or vector of quantiles
}
\description{
todo*
}
\references{
todo*
}
\author{
Justin Moat. J.Moat@kew.org
}
