% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pro_error_vol}
\alias{pro_error_vol}
\title{Volume progression of errors}
\usage{
pro_error_vol(dbh, dbhsd, height, hsd)
}
\arguments{
\item{dbh}{diameter at breast height (dbh) in cm}

\item{dbhsd}{dbh sd}

\item{height}{tree height in meters}

\item{hsd}{height sd}
}
\value{
standard error
}
\description{
Volume progression of errors
}
\note{
assumes allometric equation has no error, which is not correct...
}
\author{
Justin Moat. J.Moat@kew.org
}
