% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{c2co2e}
\alias{c2co2e}
\title{Carbon to CO2 equivalent}
\usage{
c2co2e(carbon)
}
\arguments{
\item{carbon}{carbon}
}
\value{
CO2e C02 equivalent
}
\description{
todo*
}
\references{
todo*
}
\author{
Justin Moat. J.Moat@kew.org
}
