% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_broad_tariff}
\alias{fc_broad_tariff}
\title{C board leaf tree tariff number}
\usage{
fc_broad_tariff(spcode, height, dbh)
}
\arguments{
\item{spcode}{species code}

\item{height}{tree height in metres}

\item{dbh}{diameter at breast height in cm}
}
\value{
tariff number
}
\description{
C board leaf tree tariff number
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org
}
