% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_merchtreevol}
\alias{fc_merchtreevol}
\title{Forestry merchantable volume}
\usage{
fc_merchtreevol(tariff, dbh)
}
\arguments{
\item{dbh}{diameter at breast height in cm}

\item{tarrif}{tree or stand tariff number}
}
\value{
volume m^3
}
\description{
Use the tree tariff number and DBH to estimate the mean merchantable tree volume.
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org
}
