% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_treevol}
\alias{fc_treevol}
\title{Forestry commission tree wood volume}
\usage{
fc_treevol(mtreevol, dbh)
}
\arguments{
\item{mtreevol}{merchantable tree volume}

\item{dbh}{diameter at breast height in cm (above 6.5 cm)}
}
\value{
volume m^3
}
\description{
Calculate the mean total stem volume by multiplying the mean merchantable tree volume by the coefficient from stemvol.rda
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org
}
